+++ arch-setup/03-panel.sh
@@ -1,34 +1,46 @@
 #!/bin/bash
 
-MENU="
-1 - tint2 
-2 - xfce4-panel 
-3 - polybar 
-" 
-echo $MENU
+OPTION=0
 
-read -p "Option: " OPTION
+while true; do
+  MENU="
+  1 - Xfce4-panel
+  2 - Polybar
+  3 - Wawbar
+  4 - Other
+  " 
+  echo "${MENU//[0-9]  /$'\n'}"
+  read -p "Option: " OPTION
+  
+  if [ "$OPTION" -ge 1 ] && [ "$OPTION" -le 4 ]; then
+  
+    case "$OPTION" in
+      1)
+        echo "installiung xfce4-panel and plugins..."
+        sudo pacman -S  	xfce4-panel \
+                      xfce4-datetime-plugin \
+                      xfce4-genmon-plugin \
+                      xfce4-pulseaudio-plugin  --noconfirm
+        ;;
+      2)
+        echo "installing polybar"
+        rm -rvf ~/Programs/polybar
+        sudo pacman -S polybar
+        ;;
+      3)
+        echo "installing waybar"
+        sudo pacman -S waybar
+        ;;
+      4)
+        echo "Install another panel later/"
+        ;;
+      *)
+        echo "Invalid Option"
+      esac
+      break 
 
-case "$OPTION" in
-  1)
-    echo "install tint2..."
-    yes | sudo pacman -S tint2
-    ;;
-  2)
-    echo "install xfce4-panel and plugins..."
-    sudo pacman -S  	xfce4-panel \
-                	xfce4-datetime-plugin \
-                	xfce4-genmon-plugin \
-                	xfce4-pulseaudio-plugin  --noconfirm
-    ;;
-  3)
-    echo "install dependencies..."
-    rm -rvf ~/Programs/polybar
-    git clone https://aur.archlinux.org/polybar.git ~/Programs/polybar
-    cd ~/Programs/polybar && makepkg -si
-    ;;
-  *)
-    echo "invalid value"
-    exit 1
-    ;;
-  esac
+  else
+    echo "Invalid Option, try again"
+  fi
+
+done
